assignment1/backend/requirements.txt
Flask==2.3.2
Flask-CORS==4.0.0
psycopg2-binary==2.9.6
SQLAlchemy==2.0.19
python-dotenv==1.0.0
